requirements.txt
flask
pandas
numpy
scikit-learn
imbalanced-learn
matplotlib
joblib
umap-learn
shap
gunicorn
